% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{apply_administrator_selections}
\alias{apply_administrator_selections}
\title{Apply administrator selections to Data Flow manifest}
\usage{
apply_administrator_selections(
  dataflow_manifest,
  administrator_widget_output,
  dataset_selection_module_output
)
}
\arguments{
\item{dataflow_manifest}{A data flow status manifest}

\item{administrator_widget_output}{Output from mod_administrator_widgets.R}

\item{dataset_selection_module_output}{Output from mod_dataset_selection.R}
}
\description{
Apply administrator selections to Data Flow manifest
}
