% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert_column_type}
\alias{convert_column_type}
\title{Convert date columns from date to string}
\usage{
convert_column_type(df, col_names, type)
}
\arguments{
\item{df}{a data flow status manifest.}

\item{col_names}{columns to update}

\item{type}{date, character, integer, logical, icon, or factor}
}
\description{
Convert date columns from date to string
}
