% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create a dashboard style datatable}
\usage{
create_dashboard(df, config)
}
\arguments{
\item{df}{data flow manifest}

\item{config}{Config for datatable dashboard module in \code{inst/datatable_dashboard_config.json}}
}
\description{
Create a dashboard style datatable
}
