% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.simple_cap}
\alias{.simple_cap}
\title{Convert a string so that first letter is capitalized}
\usage{
.simple_cap(x)
}
\arguments{
\item{x}{A string or vector of strings}
}
\description{
Convert a string so that first letter is capitalized
}
