% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{dt_replace_na}
\alias{dt_replace_na}
\title{NA replacement - datatable custom JS}
\usage{
dt_replace_na(col_index, na_replacement)
}
\arguments{
\item{col_index}{target columns index}

\item{na_replacement}{text to replace NA}
}
\description{
NA replacement - datatable custom JS
}
