% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{fill_dataflow_manifest}
\alias{fill_dataflow_manifest}
\title{Generate a data flow status manifest skeleton. Fills in the component, contributor, data type, number of items, and dataset name columns.}
\usage{
fill_dataflow_manifest(
  dataflow_manifest_chunk,
  schema_url,
  na_replace = NULL,
  base_url
)
}
\arguments{
\item{dataflow_manifest_chunk}{Some rows of a data flow manifest. Generally will be \code{contributor}, \code{entityId}, \code{Component}, \code{dataset}, \code{dataset_name}}

\item{schema_url}{URL of DataFlow schema jsonld (if on GitHub must be raw file)}

\item{na_replace}{String to use in place of NA. Defaults to "Not Applicable". Enter NULL for \code{NA}.}

\item{base_url}{Base URL of schematic API}
}
\description{
Generate a data flow status manifest skeleton. Fills in the component, contributor, data type, number of items, and dataset name columns.
}
