% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{generate_dashboard_config}
\alias{generate_dashboard_config}
\title{Parse config to get columns types}
\usage{
generate_dashboard_config(dcc_config, base_url)
}
\arguments{
\item{dcc_config}{DCC config}

\item{base_url}{Schematic REST API base URL}
}
\description{
Parse config to get columns types
}
