% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{generate_data_flow_manifest_skeleton}
\alias{generate_data_flow_manifest_skeleton}
\title{Generate a data flow status manifest skeleton. Fills in the component, contributor, data type, number of items, and dataset name columns.}
\usage{
generate_data_flow_manifest_skeleton(
  asset_view,
  input_token,
  calc_num_items,
  base_url = "https://schematic-dev.api.sagebionetworks.org"
)
}
\arguments{
\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{calc_num_items}{TRUE/FALSE. Calculate the number of items in each manifest.}

\item{base_url}{Base URL of schematic API}

\item{storage_project_list}{List output from \code{storage_projects} schematic endpoint}

\item{dataset_id}{Synapse ID of existing manifest}
}
\description{
Generate a data flow status manifest skeleton. Fills in the component, contributor, data type, number of items, and dataset name columns.
}
