% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{get_annotations}
\alias{get_annotations}
\title{Gets an annotation for a given entity ID}
\usage{
get_annotations(
  id,
  annotation_label,
  access_token,
  na_replace = NA,
  url = file.path("https://repo-prod.prod.sagebase.org", "repo/v1/entity")
)
}
\arguments{
\item{id}{Synapse ID}

\item{access_token}{Synapse authentication token}

\item{na_replace}{NA replacement string}

\item{url}{Synapse API endpoint URL}

\item{annotation}{label Label of annotation (case sensitive)}
}
\description{
Gets an annotation for a given entity ID
}
