% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_colname_by_type}
\alias{get_colname_by_type}
\title{Parse config to get columns types}
\usage{
get_colname_by_type(type, config)
}
\arguments{
\item{type}{column type as described in datatable_dashboard_config.json}

\item{config}{datatable_dashboard_config.json as a datatable (\code{jsonlite::read_json("inst/datatable_dashboard_config.json")})}
}
\description{
Parse config to get columns types
}
