% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_na_replace_colnames}
\alias{get_na_replace_colnames}
\title{Parse config to get columns with na_replace specified}
\usage{
get_na_replace_colnames(config)
}
\arguments{
\item{config}{datatable_dashboard_config.json as a datatable (\code{jsonlite::read_json("inst/datatable_dashboard_config.json")})}
}
\description{
Parse config to get columns with na_replace specified
}
