% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{Convert a list to a dataframe}
\usage{
list_to_dataframe(list, col_names = NULL)
}
\arguments{
\item{list}{A list of items. Only works where list entries all follow the same pattern.}

\item{col_names}{Desired column names. Ex: \code{c("id", "name")}}
}
\description{
Convert a list to a dataframe
}
