% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_download}
\alias{manifest_download}
\title{Download a manifest using the manifest Synapse ID}
\usage{
manifest_download(
  manifest_id,
  access_token,
  base_url = paste0("https://", "schematic.api.sagebionetworks.org")
)
}
\arguments{
\item{manifest_id}{Synapse ID of a manifest}

\item{access_token}{Synapse login cookie, PAT, or API key.}

\item{base_url}{URL to schematic API endpoint}
}
\description{
Download a manifest using the manifest Synapse ID
}
