% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_data_flow_status.R
\name{mod_administrator_widgets_server}
\alias{mod_administrator_widgets_server}
\title{update_data_flow_status Server Functions}
\usage{
mod_administrator_widgets_server(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
update_data_flow_status Server Functions
}
