% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_data_flow_status.R
\name{mod_administrator_widgets_ui}
\alias{mod_administrator_widgets_ui}
\title{administrator_widgets UI Function}
\usage{
mod_administrator_widgets_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
