% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard_filters.R
\name{mod_dashboard_filters_server}
\alias{mod_dashboard_filters_server}
\title{dashboard_filters Server Functions}
\usage{
mod_dashboard_filters_server(id, dashboard_config, manifest)
}
\description{
dashboard_filters Server Functions
}
