% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard_filters.R
\name{mod_dashboard_filters_ui}
\alias{mod_dashboard_filters_ui}
\title{dashboard_filters UI Function}
\usage{
mod_dashboard_filters_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
