% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset_selection.R
\name{mod_dataset_selection_ui}
\alias{mod_dataset_selection_ui}
\title{Select a Dataset UI Function}
\usage{
mod_dataset_selection_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A Shiny module that displays the datasets in a given asset view in a selectable DT::datatable
}
