% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datatable_dashboard.R
\name{mod_datatable_dashboard_server}
\alias{mod_datatable_dashboard_server}
\title{datatable_dashboard Server Functions}
\usage{
mod_datatable_dashboard_server(id, df, config)
}
\arguments{
\item{id}{shiny id}

\item{df}{dataframe containing data to be displayed in dashboard}

\item{config}{config in \code{inst/datatable_dashboard_config.json}}
}
\description{
datatable_dashboard Server Functions
}
