% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datatable_dashboard.R
\name{mod_datatable_dashboard_ui}
\alias{mod_datatable_dashboard_ui}
\title{Datatable Dashboard UI Function}
\usage{
mod_datatable_dashboard_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny module that displays a DataFlow manifest as a dashboard
}
