% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datatable_filters.R
\name{mod_datatable_filters_server}
\alias{mod_datatable_filters_server}
\title{Filters for dataFlow Manifest Server.}
\usage{
mod_datatable_filters_server(id, manifest)
}
\arguments{
\item{id}{internal Shiny ID}

\item{manifest}{a dataFlow manifest}
}
\value{
a filtered dataframe
}
\description{
Filters for dataFlow Manifest Server.
}
