% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribution.R
\name{mod_distribution_server}
\alias{mod_distribution_server}
\title{Distribution Plot Server Functions}
\usage{
mod_distribution_server(
  id,
  df,
  group_by_var,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  fill = "#0d1c38"
)
}
\arguments{
\item{id}{Shiny ID to call server module}

\item{df}{A data frame containing data to plot}

\item{group_by_var}{Column variable to group data by (x axis variable)}

\item{title}{Title of plot}

\item{x_lab}{X axis label}

\item{y_lab}{Y axis label}

\item{fill}{Fill color}
}
\description{
Distribution Plot Server Functions
}
