% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribution.R
\name{mod_distribution_ui}
\alias{mod_distribution_ui}
\title{Distribution Plot UI Function}
\usage{
mod_distribution_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A Shiny module that takes a dataframe and displays a distribution bar plot
}
