% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_selection.R
\name{mod_file_selection_server}
\alias{mod_file_selection_server}
\title{Select a File Server Functions}
\usage{
mod_file_selection_server(id, dataset, asset_view, input_token)
}
\arguments{
\item{id}{internal shiny ID}

\item{dataset}{dataset folder synId}

\item{asset_view}{fileview synId}

\item{input_token}{Syanpse PAT}
}
\description{
Select a File Server Functions
}
