% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_selection.R
\name{mod_file_selection_ui}
\alias{mod_file_selection_ui}
\title{Select a File UI function}
\usage{
mod_file_selection_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A Shiny module that displays the files in a given dataset in a selectable DT::datatable
}
