% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_highlight_datatable.R
\name{mod_highlight_datatable_server}
\alias{mod_highlight_datatable_server}
\title{Datatable that highlights based on a given input Server}
\usage{
mod_highlight_datatable_server(id, df, selection, df_match_colname)
}
\arguments{
\item{id}{internal Shiny ID}

\item{df}{a dataframe}

\item{selection}{a vector of IDs to match on}

\item{df_match_columns}{column name to match on}
}
\description{
Datatable that highlights based on a given input Server
}
