% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_highlight_datatable.R
\name{mod_highlight_datatable_ui}
\alias{mod_highlight_datatable_ui}
\title{Datatable that highlights based on a given input UI}
\usage{
mod_highlight_datatable_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
