% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manifest_preview.R
\name{mod_manifest_preview_server}
\alias{mod_manifest_preview_server}
\title{Manifest Preview Server}
\usage{
mod_manifest_preview_server(id, df, selection, df_match_colname)
}
\arguments{
\item{id}{shiny id}

\item{df}{a dataframe}

\item{selection}{a vector of IDs to highlight}

\item{df_match_colname}{column name to match highlight on}
}
\description{
Shows preview of manifest that highlights based on a given input
}
