% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manifest_preview.R
\name{mod_manifest_preview_ui}
\alias{mod_manifest_preview_ui}
\title{Manifest Preview UI}
\usage{
mod_manifest_preview_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Shows preview of manifest that highlights based on a given input
}
