% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scheduler.R
\name{mod_scheduler_server}
\alias{mod_scheduler_server}
\title{scheduler Server Functions}
\usage{
mod_scheduler_server(id, reset_btn)
}
\arguments{
\item{id}{shiny id}

\item{reset_btn}{reset button input}
}
\description{
scheduler Server Functions
}
