% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scheduler.R
\name{mod_scheduler_ui}
\alias{mod_scheduler_ui}
\title{scheduler UI Function}
\usage{
mod_scheduler_ui(id, dateInput_label, checkboxInput_label)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{dateInput_label}{label for dateInput}

\item{checkboxInput_label}{date for checkboxInput}
}
\description{
A shiny Module.
}
