% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_storage_project.R
\name{mod_select_storage_project_server}
\alias{mod_select_storage_project_server}
\title{Storage Project Selection Module Server}
\usage{
mod_select_storage_project_server(
  id,
  asset_view,
  access_token,
  base_url = paste0("https://", "schematic.api.sagebionetworks", ".org")
)
}
\arguments{
\item{id}{shiny id}

\item{asset_view}{An ID of a file view listing all project data assets}

\item{access_token}{A Synapse Personal Access Token}

\item{base_url}{A Schematic API base URL. Default is
\verb{https://schematic.api.sagebionetworks.org}}
}
\description{
Storage Project Selection Module Server
}
