% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_storage_project.R
\name{mod_select_storage_project_ui}
\alias{mod_select_storage_project_ui}
\title{Storage Project Selection Module UI}
\usage{
mod_select_storage_project_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny module. Outputs a selectInput dropdown of Synapse storage project names to the UI.
}
