% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stacked_bar.R
\name{mod_stacked_bar_server}
\alias{mod_stacked_bar_server}
\title{Stacked Bar Plot Server Functions}
\usage{
mod_stacked_bar_server(
  id,
  df,
  x_var,
  y_var,
  fill_var,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  colors = NULL,
  x_line = NULL,
  width = NULL,
  date_breaks = NULL,
  coord_flip = FALSE
)
}
\arguments{
\item{id}{Shiny ID to call server module}

\item{df}{A data frame containing data to plot}

\item{x_var}{Column name of X variable}

\item{y_var}{Column name of Y variable}

\item{fill_var}{Column name of fill variable}

\item{title}{Title of plot}

\item{x_lab}{X axis label}

\item{y_lab}{Y axis label}

\item{colors}{Fill colors}

\item{x_line}{Add a line across x axis}

\item{width}{Width of stacked bars}

\item{date_breaks}{X axis date breaks}

\item{coord_flip}{Flip coordinates 90 degrees}
}
\description{
Stacked Bar Plot Server Functions
}
