% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_submit_model.R
\name{mod_submit_model_ui}
\alias{mod_submit_model_ui}
\title{submit_model UI Function}
\usage{
mod_submit_model_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
