% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tabbed_dashboard.R
\name{mod_tabbed_dashboard_server}
\alias{mod_tabbed_dashboard_server}
\title{Tabbed dashboard module UI (deprecated)}
\usage{
mod_tabbed_dashboard_server(id, df, config)
}
\arguments{
\item{id}{tabbed dashboard ID}

\item{df}{dataframe with certain columns}

\item{config}{datatable dashboard config}
}
\description{
Tabbed dashboard module UI (deprecated)
}
