% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tabbed_dashboard.R
\name{mod_tabbed_dashboard_ui}
\alias{mod_tabbed_dashboard_ui}
\title{Tabbed dashboard module UI (deprecated)}
\usage{
mod_tabbed_dashboard_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny module that displays a tabbed dashboard
}
