% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_data_flow_status.R
\name{mod_update_data_flow_status_server}
\alias{mod_update_data_flow_status_server}
\title{update_data_flow_status Server Functions}
\usage{
mod_update_data_flow_status_server(id)
}
\description{
update_data_flow_status Server Functions
}
