% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_data_flow_status.R
\name{mod_update_data_flow_status_ui}
\alias{mod_update_data_flow_status_ui}
\title{update_data_flow_status UI Function}
\usage{
mod_update_data_flow_status_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
