% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datatable_filters.R
\name{\%modifiedIn\%}
\alias{\%modifiedIn\%}
\title{Filtering operator}
\usage{
e1 \%modifiedIn\% e2
}
\arguments{
\item{e1}{function will check if the values of the first argument are present in the second argument}

\item{e2}{function will check if the values of the first argument are present in the second argument}
}
\description{
A special \%in\% operator that doesn't filter nulls (returns TRUE)
}
