% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{prep_manifest_dash}
\alias{prep_manifest_dash}
\title{Prepare a dataframe for a dashboard style datatable}
\usage{
prep_manifest_dash(manifest, config)
}
\arguments{
\item{manifest}{data flow manifest}

\item{config}{Config for datatable dashboard module in \code{inst/datatable_dashboard_config.json}}
}
\description{
Prepare a dataframe for a dashboard style datatable
}
