% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{prep_manifest_dfa}
\alias{prep_manifest_dfa}
\title{Prepare a dataframe that has been downloaded from Synapse for the Data Flow App}
\usage{
prep_manifest_dfa(manifest, config)
}
\arguments{
\item{manifest}{A manifest that has been downloaded from using manifest_download_to_df()}

\item{config}{\code{datatable_dashboard_config.json} read in as a dataframe}
}
\description{
Prepare a dataframe that has been downloaded from Synapse for the Data Flow App
}
