% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rearrange_dataframe}
\alias{rearrange_dataframe}
\title{Rearrange a dataframe. Expected columns are ordered to match \code{expected_col_names} and moved to the front of the dataframe. Unexpected columns are moved to the back of the dataframe.}
\usage{
rearrange_dataframe(df, expected_col_names)
}
\arguments{
\item{df}{a dataframe}

\item{expected_col_names}{a vector of expected column names in the order you would like them to appear.}
}
\description{
Rearrange a dataframe. Expected columns are ordered to match \code{expected_col_names} and moved to the front of the dataframe. Unexpected columns are moved to the back of the dataframe.
}
