% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{schemas_get_node_validation_rules}
\alias{schemas_get_node_validation_rules}
\title{Get all the attributes associated with a specific data model component
formatted as a dataframe}
\usage{
schemas_get_node_validation_rules(schema_url, node_display_name, base_url)
}
\arguments{
\item{schema_url}{A data model URL}

\item{node_display_name}{Display lable of node}

\item{base_url}{URL to schematic API endpoint}
}
\description{
Get all the attributes associated with a specific data model component
formatted as a dataframe
}
