% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{storage_project_datasets}
\alias{storage_project_datasets}
\title{Gets all datasets in a given storage project}
\usage{
storage_project_datasets(
  asset_view,
  project_id,
  access_token,
  base_url = paste0("https://", "schematic.api.", "sagebionetworks.org")
)
}
\arguments{
\item{asset_view}{synapse ID of master file view.}

\item{project_id}{synapse ID of a storage project.}

\item{access_token}{synapse PAT}

\item{base_url}{URL to schematic API endpoint}
}
\description{
Gets all datasets in a given storage project
}
