% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{storage_project_manifests}
\alias{storage_project_manifests}
\title{/storage/project/manifests}
\usage{
storage_project_manifests(
  asset_view,
  project_id,
  input_token,
  base_url = "https://schematic-dev.api.sagebionetworks.org"
)
}
\arguments{
\item{asset_view}{synapse ID of master file view.}

\item{input_token}{synapse PAT}

\item{base_url}{URL to schematic API endpoint}

\item{dataset_id}{synapse ID of a storage dataset.}
}
\description{
/storage/project/manifests
}
