% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{storage_projects}
\alias{storage_projects}
\title{Get all storage projects the current user has access to}
\usage{
storage_projects(
  asset_view,
  access_token,
  base_url = "https://schematic.api.sagebionetworks.org"
)
}
\arguments{
\item{asset_view}{synapse ID of master file view.}

\item{access_token}{synapse PAT}

\item{base_url}{URL to schematic API endpoint}
}
\description{
Get all storage projects the current user has access to
}
