% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{style_dashboard}
\alias{style_dashboard}
\title{Add custom styling to dashboard based on contents of config}
\usage{
style_dashboard(prepped_manifest, config)
}
\arguments{
\item{prepped_manifest}{A manifest prepped with \code{prep_manifest_dash()}}

\item{config}{\code{inst/datatable_dashboard_config.json}}
}
\description{
Add custom styling to dashboard based on contents of config
}
