% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_annotations}
\alias{synapse_annotations}
\title{Gets annotations for a given entity ID}
\usage{
synapse_annotations(
  id,
  auth,
  url = "https://repo-prod.prod.sagebase.org/repo/v1/entity"
)
}
\arguments{
\item{id}{Synapse ID}

\item{auth}{Synapse authentication token}

\item{url}{URL to REST API endpoint}
}
\description{
wrapper for https://rest-docs.synapse.org/rest/GET/entity/id/annotations2.html
}
