% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_permissions}
\alias{synapse_permissions}
\title{Gets user permissions}
\usage{
synapse_permissions(
  id,
  auth,
  url = "https://repo-prod.prod.sagebase.org/repo/v1/entity"
)
}
\arguments{
\item{id}{Synapse ID}

\item{auth}{Synapse authentication token}

\item{url}{Synapse API endpoint URL}
}
\description{
Gets user permissions
}
