% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{true_false_icon}
\alias{true_false_icon}
\title{Convert a vector of TRUE/FALSE to icon html}
\usage{
true_false_icon(vec)
}
\arguments{
\item{vec}{A vector of TRUE/FALSE}
}
\description{
Convert a vector of TRUE/FALSE to icon html
}
