% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{update_data_flow_manifest}
\alias{update_data_flow_manifest}
\title{Check synapse for updates to data flow status manifest}
\usage{
update_data_flow_manifest(
  asset_view,
  manifest_dataset_id,
  input_token,
  base_url
)
}
\arguments{
\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{manifest_dataset_id}{Dataset ID for data flow status manifest to be updated}

\item{input_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API (Defaults to  AWS version)}
}
\description{
Check synapse for updates to data flow status manifest
}
