% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{update_dfs_manifest}
\alias{update_dfs_manifest}
\title{Convert a list to a dataframe}
\usage{
update_dfs_manifest(dfs_manifest, dfs_updates, selected_datasets_df)
}
\arguments{
\item{dfs_manifest}{A data flow status manifest}

\item{dfs_updates}{Output from mod_update_data_flow_status.R}

\item{selected_datasets_df}{Output from mod_dataset_selection.R}
}
\description{
Convert a list to a dataframe
}
