% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{update_manifest_add_datasets}
\alias{update_manifest_add_datasets}
\title{Update manifest with new datasets found in Synapse}
\usage{
update_manifest_add_datasets(
  dataflow_manifest,
  get_all_manifests_out,
  asset_view,
  na_replace,
  schema_url,
  access_token,
  base_url
)
}
\arguments{
\item{dataflow_manifest}{A dataFlow manifest}

\item{get_all_manifests_out}{The output of get_all_manifests.
Also can be a dataframe that includes Component, contributor,
entityId, dataset_name, and dataset.}

\item{asset_view}{ID of view listing all project data assets}

\item{na_replace}{NA replacement text}

\item{schema_url}{A data model URL}

\item{access_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API}
}
\description{
Update manifest with new datasets found in Synapse
}
