% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{update_manifest_column}
\alias{update_manifest_column}
\title{Update dataFlow manifest when dataset folder name changes}
\usage{
update_manifest_column(
  dataflow_manifest,
  get_all_manifests_out,
  update_column,
  asset_view,
  recalc_num_items = FALSE,
  input_token,
  base_url
)
}
\arguments{
\item{dataflow_manifest}{A dataFlow manifest}

\item{get_all_manifests_out}{The output of get_all_manifests. Also can be a dataframe that includes Component, contributor, entityId, dataset_name, and dataset.}

\item{update_column}{Column name of the column to be updated}

\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{recalc_num_items}{TRUE/FALSE if there is an item to be updated, should the manifest}

\item{input_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API (Defaults to  AWS version)}
}
\description{
Update dataFlow manifest when dataset folder name changes
}
