% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{update_manifest_remove_datasets}
\alias{update_manifest_remove_datasets}
\title{Remove datasets that are no longer found in Synapse}
\usage{
update_manifest_remove_datasets(
  dataflow_manifest,
  get_all_manifests_out,
  asset_view,
  access_token,
  base_url
)
}
\arguments{
\item{dataflow_manifest}{A dataFlow manifest}

\item{get_all_manifests_out}{The output of get_all_manifests. Also can be a dataframe that includes Component, contributor, entityId, dataset_name, and dataset.}

\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{access_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API (Defaults to  AWS version)}
}
\description{
Remove datasets that are no longer found in Synapse
}
