% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{visualize_component}
\alias{visualize_component}
\title{Get all the attributes associated with a specific data model component
formatted as a dataframe}
\usage{
visualize_component(
  schema_url,
  component = "DataFlow",
  base_url = "https://schematic.api.sagebionetworks.org"
)
}
\arguments{
\item{schema_url}{A data model URL}

\item{component}{Component of the data model to explore}

\item{base_url}{URL to schematic API endpoint}
}
\description{
Get all the attributes associated with a specific data model component
formatted as a dataframe
}
