% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{calculate_items_per_manifest}
\alias{calculate_items_per_manifest}
\title{Call \code{calculate_items_per_manifest} calculate the number of items per manifest synID in a given dataframe}
\usage{
calculate_items_per_manifest(df, asset_view, access_token, base_url)
}
\arguments{
\item{df}{A dataframe with \code{entityId} and \code{dataset} columns from data flow status manifest}

\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{access_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API}
}
\description{
Call \code{calculate_items_per_manifest} calculate the number of items per manifest synID in a given dataframe
}
