% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{dt_replace_na}
\alias{dt_replace_na}
\title{Dynamically create custom NA replacement text}
\usage{
dt_replace_na(col_index, na_replacement)
}
\arguments{
\item{col_index}{target column index}

\item{na_replacement}{NA replacement string}
}
\description{
Dynamically create custom NA replacement text
}
