% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{generate_dashboard_config}
\alias{generate_dashboard_config}
\title{Parse config to get columns types}
\usage{
generate_dashboard_config(
  schema_url,
  display_names = NULL,
  icon = TRUE,
  na_replace = NULL,
  base_url
)
}
\arguments{
\item{schema_url}{URL of a Schematic data model schema}

\item{display_names}{Vector of display names to use in dashboard}

\item{icon}{Display \code{Valid_Values = TRUE/FALSE} attributes as icons in dashboard}

\item{na_replace}{Named list indicating strings to replace NA with \code{na_replace <- list(attribute_1 = "na string 1", attribute_2 = "na string 2")}}

\item{base_url}{Schematic REST API base URL}
}
\description{
Parse config to get columns types
}
