% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_download}
\alias{manifest_download}
\title{Download a manifest}
\usage{
manifest_download(
  asset_view,
  dataset_id,
  input_token,
  base_url = "https://schematic-dev.api.sagebionetworks.org"
)
}
\arguments{
\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{dataset_id}{Synapse ID of existing manifest}

\item{input_token}{Synapse login cookie, PAT, or API key.}

\item{base_url}{URL to schematic API endpoint}
}
\description{
Download a manifest
}
