% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datatable_filters.R
\name{mod_datatable_filters_ui}
\alias{mod_datatable_filters_ui}
\title{Filters for dataFlow Manifest UI}
\usage{
mod_datatable_filters_ui(id, width = NULL)
}
\arguments{
\item{id}{shiny id}

\item{width}{filter width}
}
\description{
A shiny Module that renders filters for the data flow manifest.
}
