% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dcc.R
\name{mod_select_dcc_server}
\alias{mod_select_dcc_server}
\title{Select a DCC Module Server}
\usage{
mod_select_dcc_server(id, tenants_config_path, access_token)
}
\arguments{
\item{id}{module ID}

\item{tenants_config_path}{DCC configuration file sourced from
\code{Sage-Bionetworks/data_flow_config}}

\item{access_token}{Synapse PAT}
}
\description{
Select a DCC Module Server
}
