% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dcc.R
\name{mod_select_dcc_ui}
\alias{mod_select_dcc_ui}
\title{Select a DCC UI}
\usage{
mod_select_dcc_ui(id, title = "Select a DCC")
}
\arguments{
\item{id}{Module id}

\item{title}{Title of box element}
}
\description{
A shiny module. Outputs a selectInput dropdown of Synapse
storage project names to the UI.
}
