% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stacked_bar.R
\name{mod_stacked_bar_ui}
\alias{mod_stacked_bar_ui}
\title{Stacked Bar Plot UI Function}
\usage{
mod_stacked_bar_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A Shiny module that takes in a datafame and outputs a stacked bar plot.
}
