% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_submit_model.R
\name{mod_submit_model_server}
\alias{mod_submit_model_server}
\title{submit_model Server Functions}
\usage{
mod_submit_model_server(
  id,
  dfs_manifest,
  data_type,
  asset_view,
  dataset_id,
  manifest_dir = "./manifest",
  input_token,
  schema_url,
  base_url
)
}
\description{
submit_model Server Functions
}
