% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{verbose_message}
\alias{verbose_message}
\title{Convert a string so that first letter is capitalized}
\usage{
verbose_message(m, verbose, type = "print")
}
\arguments{
\item{m}{a string}

\item{verbose}{TRUE/FALSE}

\item{type}{Type of message to surface (print or message)}
}
\description{
Convert a string so that first letter is capitalized
}
